requirements.txt
asgiref==3.8.1
Django==5.1.4
django-recaptcha==4.0.0
django-tinymce==4.1.0
pillow==11.0.0
six==1.17.0
sqlparse==0.5.3
